requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
djangorestframework_simplejwt
twilio
drf-yasg
django-filter
gunicorn
python-decouple
celery
redis

